// src/api/generated/model/createOrdContractIssuanceRuleRequestDtoSourceType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип источника аудитории
 */
export type CreateOrdContractIssuanceRuleRequestDtoSourceType = typeof CreateOrdContractIssuanceRuleRequestDtoSourceType[keyof typeof CreateOrdContractIssuanceRuleRequestDtoSourceType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateOrdContractIssuanceRuleRequestDtoSourceType = {
  room: 'room',
  event: 'event',
  creativeTask: 'creativeTask',
} as const;